pyproject.toml
[tool.poetry]
name = "api-django-ninja"
version = "0.1.0"
description = ""
authors = ["Edevaldo <edevaldohbfilho@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
django-ninja = "^1.3.0"
django-environ = "^0.11.2"
psycopg2 = "^2.9.9"
python-dotenv = "^1.0.1"
cpf-cnpj-validate = "^1.4"


[tool.poetry.group.dev.dependencies]
commitizen = "^3.29.0"
pre-commit = "^3.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
